+++ src/Quartz.Tests.Unit/Extensions/DependencyInjection/MultipleSchedulerTests.cs
@@ -1,4 +1,4 @@
-using System.Collections.Specialized;
+﻿using System.Collections.Specialized;
 
 
 using Microsoft.Extensions.Configuration;
@@ -429,6 +429,8 @@ public sealed class MultipleSchedulerTests
 
     private sealed class TestSchedulerListenerA : ISchedulerListener
     {
+        public string Name => nameof(TestSchedulerListenerA);
+
         public ValueTask JobScheduled(ITrigger trigger, CancellationToken cancellationToken = default) => default;
         public ValueTask JobUnscheduled(TriggerKey triggerKey, CancellationToken cancellationToken = default) => default;
         public ValueTask TriggerFinalized(ITrigger trigger, CancellationToken cancellationToken = default) => default;
@@ -442,12 +444,12 @@ public sealed class MultipleSchedulerTests
         public ValueTask JobsPaused(string jobGroup, CancellationToken cancellationToken = default) => default;
         public ValueTask JobResumed(JobKey jobKey, CancellationToken cancellationToken = default) => default;
         public ValueTask JobsResumed(string jobGroup, CancellationToken cancellationToken = default) => default;
-        public ValueTask SchedulerError(string msg, SchedulerException cause, CancellationToken cancellationToken = default) => default;
+        public ValueTask SchedulerError(string message, SchedulerException exception, CancellationToken cancellationToken = default) => default;
         public ValueTask SchedulerInStandbyMode(CancellationToken cancellationToken = default) => default;
         public ValueTask SchedulerStarted(CancellationToken cancellationToken = default) => default;
         public ValueTask SchedulerStarting(CancellationToken cancellationToken = default) => default;
         public ValueTask SchedulerShutdown(CancellationToken cancellationToken = default) => default;
-        public ValueTask SchedulerShuttingdown(CancellationToken cancellationToken = default) => default;
+        public ValueTask SchedulerShuttingDown(CancellationToken cancellationToken = default) => default;
         public ValueTask SchedulingDataCleared(CancellationToken cancellationToken = default) => default;
         public ValueTask JobInterrupted(JobKey jobKey, CancellationToken cancellationToken = default) => default;
     }